+++ libs/qwt/src/qwt_plot_spectrocurve.cpp
@@ -237,7 +237,7 @@ double QwtPlotSpectroCurve::penWidth() const
  */
 void QwtPlotSpectroCurve::drawSeries(QPainter* painter,
     const QwtScaleMap& xMap, const QwtScaleMap& yMap,
-    const QRectF& canvasRect, size_t from, size_t to ) const
+    const QRectF& canvasRect, int from, int to ) const
 {
     if ( !painter || dataSize() <= 0 )
         return;
